Checkpoints working

+++ app/Services/TaskService.js
@@ -15,39 +15,38 @@ async addTask(taskData){
 let newTask = new Task(taskData)
 let res = await taskApi.post('', newTask)
 ProxyState.tasks = [...ProxyState.tasks, new Task(res.data)]
-ProxyState.taskCount ++
 console.log(res.data)
-console.log(ProxyState.taskCount)
 
 }
 
 async getTasks(){
   let res = await taskApi.get()
   ProxyState.tasks = res.data.map(t => new Task(t))
-  ProxyState.taskCount = ProxyState.tasks.length
   console.log('is it working?', ProxyState.tasks)
-  console.log(ProxyState.taskCount)
 }
 
-async removeTask(id){
-  await taskApi.delete(id)
-  ProxyState.tasks = ProxyState.tasks.filter(t => t.id !== id)
-  console.log(ProxyState.taskCount)
-}
 
 async finishedTask(Taskid){
   let task = await ProxyState.tasks.find(t => t.id == Taskid);
-
+  
   if(task.completed){
     task.completed = false
-    ProxyState.completed -= 1
   } else {
     task.completed = true
-    ProxyState.completed += 1
   }
 let res = await taskApi.put(Taskid, task)
+console.log('is this triggering')
 this.getTasks()
-  
+
+
+}
+async removeTask(id){
+  let yes = confirm("Are you sure you want to delete this Todo?")
+  if(yes == true)
+  await taskApi.delete(id)
+this.getTasks()
+// ProxyState.tasks = ProxyState.tasks.filter(t => t.id !== id)
+
 }
 
 }
